Add scorer

+++ Runner/Runner/Writer.swift
@@ -11,7 +11,7 @@ import Foundation
 struct Writer {
     let solutionFolder = ProcessInfo.processInfo.environment["SOLUTIONS_PATH"]!
     
-    func write(solution: String, solverName: String, score: UInt64, problemName: String) throws {
+    func write(solution: String, solverName: String, score: Int, problemName: String) throws {
         let name = "\(problemName)_\(score)_\(solverName).out"
         let documentsUrl =  URL(fileURLWithPath: solutionFolder, isDirectory: true)
         let file = documentsUrl.appendingPathComponent(name)
